requirements.txt
django=3.0.2
mysql-connector-python
django-rest-swagger
django-cors-headers
mysqlclient
djangorestframework
markdown
django-filter
pyyaml
ua-parser 
user-agents
django-user-agents